+++ parabolic/src/parabolic.h
@@ -6,64 +6,64 @@
 typedef struct {
 
   // x1 bdry1
-  double *dif_b1_x_k_loc; 
-  double *dif_b1_y_k_loc; 
-  double *dif_b1_z_k_loc; 
-  double *dif_b1_x_k1_loc;
-  double *dif_b1_y_k1_loc;
-  double *dif_b1_z_k1_loc;
+  float *dif_b1_x_k_loc; 
+  float *dif_b1_y_k_loc; 
+  float *dif_b1_z_k_loc; 
+  float *dif_b1_x_k1_loc;
+  float *dif_b1_y_k1_loc;
+  float *dif_b1_z_k1_loc;
 
-  double *dif_b1_x_k; 
-  double *dif_b1_y_k; 
-  double *dif_b1_z_k; 
-  double *dif_b1_x_k1;
-  double *dif_b1_y_k1;
-  double *dif_b1_z_k1;
+  float *dif_b1_x_k; 
+  float *dif_b1_y_k; 
+  float *dif_b1_z_k; 
+  float *dif_b1_x_k1;
+  float *dif_b1_y_k1;
+  float *dif_b1_z_k1;
 
   // x2 bdry2
-  double *dif_b2_x_k_loc; 
-  double *dif_b2_y_k_loc; 
-  double *dif_b2_z_k_loc; 
-  double *dif_b2_x_k1_loc;
-  double *dif_b2_y_k1_loc;
-  double *dif_b2_z_k1_loc;
+  float *dif_b2_x_k_loc; 
+  float *dif_b2_y_k_loc; 
+  float *dif_b2_z_k_loc; 
+  float *dif_b2_x_k1_loc;
+  float *dif_b2_y_k1_loc;
+  float *dif_b2_z_k1_loc;
 
-  double *dif_b2_x_k; 
-  double *dif_b2_y_k; 
-  double *dif_b2_z_k; 
-  double *dif_b2_x_k1;
-  double *dif_b2_y_k1;
-  double *dif_b2_z_k1;
+  float *dif_b2_x_k; 
+  float *dif_b2_y_k; 
+  float *dif_b2_z_k; 
+  float *dif_b2_x_k1;
+  float *dif_b2_y_k1;
+  float *dif_b2_z_k1;
 
   // y1 bdry3
-  double *dif_b3_x_k_loc; 
-  double *dif_b3_y_k_loc; 
-  double *dif_b3_z_k_loc; 
-  double *dif_b3_x_k1_loc;
-  double *dif_b3_y_k1_loc;
-  double *dif_b3_z_k1_loc;
+  float *dif_b3_x_k_loc; 
+  float *dif_b3_y_k_loc; 
+  float *dif_b3_z_k_loc; 
+  float *dif_b3_x_k1_loc;
+  float *dif_b3_y_k1_loc;
+  float *dif_b3_z_k1_loc;
 
-  double *dif_b3_x_k; 
-  double *dif_b3_y_k; 
-  double *dif_b3_z_k; 
-  double *dif_b3_x_k1;
-  double *dif_b3_y_k1;
-  double *dif_b3_z_k1;
+  float *dif_b3_x_k; 
+  float *dif_b3_y_k; 
+  float *dif_b3_z_k; 
+  float *dif_b3_x_k1;
+  float *dif_b3_y_k1;
+  float *dif_b3_z_k1;
 
   // y2 bdry4
-  double *dif_b4_x_k_loc; 
-  double *dif_b4_y_k_loc; 
-  double *dif_b4_z_k_loc; 
-  double *dif_b4_x_k1_loc;
-  double *dif_b4_y_k1_loc;
-  double *dif_b4_z_k1_loc;
+  float *dif_b4_x_k_loc; 
+  float *dif_b4_y_k_loc; 
+  float *dif_b4_z_k_loc; 
+  float *dif_b4_x_k1_loc;
+  float *dif_b4_y_k1_loc;
+  float *dif_b4_z_k1_loc;
 
-  double *dif_b4_x_k; 
-  double *dif_b4_y_k; 
-  double *dif_b4_z_k; 
-  double *dif_b4_x_k1;
-  double *dif_b4_y_k1;
-  double *dif_b4_z_k1;
+  float *dif_b4_x_k; 
+  float *dif_b4_y_k; 
+  float *dif_b4_z_k; 
+  float *dif_b4_x_k1;
+  float *dif_b4_y_k1;
+  float *dif_b4_z_k1;
 
 } bdry_effct_t;
 /*************************************************
@@ -76,23 +76,23 @@ para_gene(gd_t *gdcurv, mympi_t *mympi,
 
 int 
 predict_point(gd_t *gdcurv, bdry_effct_t *bdry_effct, mympi_t *mympi,
-              int k, int o2i, double coef, double *x1_len,
-              double *x2_len, double *y1_len, double *y2_len);
+              int k, int o2i, float coef, float *x1_len,
+              float *x2_len, float *y1_len, float *y2_len);
 
 int
-update_point(gd_t *gdcurv, double *var_th, int k, double *coord); 
+update_point(gd_t *gdcurv, float *var_th, int k, float *coord); 
 
 int
 bdry_modify(gd_t *gdcurv, bdry_effct_t *bdry_effct, mympi_t *mympi, int k);
 
 int
-flip_bdry_z(double *x1, double *x2, double *y1, double *y2,
+flip_bdry_z(float *x1, float *x2, float *y1, float *y2,
             int total_nx, int total_ny, int total_nz);
 
 int
-cal_bdry_arc_length(double *x1, double *x2, double *y1, double *y2, 
+cal_bdry_arc_length(float *x1, float *x2, float *y1, float *y2, 
                     int total_nx, int total_ny, int total_nz, 
-                    double *x1_len, double *x2_len, double *y1_len, double *y2_len);
+                    float *x1_len, float *x2_len, float *y1_len, float *y2_len);
 
 int
 exchange_coord(gd_t *gdcurv, mympi_t *mympi, int k, int num_of_s_reqs, int num_of_r_reqs);
